2023년 6월  4일 일요일 22시 00분 40초 KST

+++ ft_atoi_pro.c
@@ -39,8 +39,8 @@ void	my_check_limit(long long num)
 
 int	ft_atoi_pro(const char *str)
 {
-	int	sign;
-	int	result;
+	int			sign;
+	long long	result;
 
 	sign = 1;
 	result = 0;
